requirements.txt
numpy==1.26.4
pillow
onnxruntime-gpu==1.18.1
opencv-contrib-python